Update Statistics component to support distributed deployment
