+++ components/org.wso2.carbon.stream.processor.statistics/src/gen/java/org/wso2/carbon/stream/processor/statistics/api/SystemDetailsApi.java
@@ -21,7 +21,12 @@ package org.wso2.carbon.stream.processor.statistics.api;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.osgi.service.component.annotations.*;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
+import org.osgi.service.component.annotations.ReferencePolicy;
 import org.wso2.carbon.analytics.msf4j.interceptor.common.AuthenticationInterceptor;
 import org.wso2.carbon.stream.processor.statistics.factories.SystemDetailsApiServiceFactory;
 import org.wso2.carbon.stream.processor.statistics.internal.service.ConfigServiceComponent;
